chore(loc): split the two files the base layer pushed over the 500-line cap
Neither file is over on `main` — the base layer grows both past the cap, and
neither is allow-listed, so the check fails for this PR and everything
stacked on it. Both splits are pure moves: no logic, no signatures, no
behaviour touched.

  weight_loader/dflash_loader.rs   575 -> 454  (+ dflash_loader/config.rs 134)
  examples/batchm_bench.rs         504 -> 461  (+ batchm_bench/dtype.rs     55)

`dflash_loader.rs` splits at the seam between DESCRIBING the checkpoint and
LOADING it: the `serde` config shapes (`DflashConfig`, `DflashRopeScaling`,
`DflashSubConfig` and their defaults) move to `dflash_loader/config.rs`,
re-exported from the parent so every existing path still resolves. Weight
structures and the loader stay put. Follows the `laguna.rs` + `laguna/`
convention already in this directory.

`batchm_bench.rs` moves its scalar dtype helpers and the deterministic RNG
(`XorShift`, the BF16 / E2M1 / E4M3 conversions the CPU-side correctness gate
compares against) to `batchm_bench/dtype.rs`, reached by `#[path]` so Cargo
does not treat it as a second example target.

Done here rather than in the base PR because the stack cannot go green
otherwise; it will conflict with further edits to those two files, and the
resolution is "take both sides" since nothing moved within the blocks.

(cherry picked from commit c4f8031de797e7213e970fa91e8baa4c83481a4e)
